[Fix](grace-exit) Fix threadpool hang on grace exit (#33132)
Fix threadpool hang on grace exit
